+++ app/src/main/java/com/example/weatherapp/holder/CityViewHolder.kt
@@ -3,5 +3,9 @@ package com.example.weatherapp.holder
 import androidx.recyclerview.widget.RecyclerView
 import com.example.weatherapp.databinding.ItemCityNameBinding
 
-class CityViewHolder(val binding: ItemCityNameBinding) :
-    RecyclerView.ViewHolder(binding.root)
+class CityViewHolder(private val binding: ItemCityNameBinding) :
+    RecyclerView.ViewHolder(binding.root) {
+    fun onBind(item: String) {
+        binding.city.text = item
+    }
+}
